# swagger.yaml
openapi: 3.0.0
info:
    version: 1.0.0
    title: MatchUp
    description: BackEnd for the wbd project
...
paths:
...
    /users/{userID}:
        get:
            summary: Get a user by id
            tags: [User]
            parameters:
                - name: userID
                  in: path
                  required: true
                  description: Autogenerated User ID by Mongoose
                  schema:
                      type: string
            responses:
                "200":
                    description: Successfully Fetched single product by the id.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/User"
        patch:
            summary: Update a User by ID
            tags: [User]
            parameters:
                - name: userID
                  in: path
                  required: true
                  description: Autogenerated User ID by Mongoose
                  schema:
                      type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/User"
            responses:
                "200":
                    description: Update a single User by the id.
                    content:
                        application/json:
                            schema: 
                                $ref: "#/components/schemas/User"
        delete:
            summary: Delete a user by id
            tags: [User]
            parameters:
                - name: userID
                  in: path
                  required: true
                  description: Autogenerated User ID by Mongoose
                  schema:
                      type: string
            responses:
                "200":
                    description: Successfully Deleted User by the id.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/User"
    /users/signin :
        post:
            summary: User Signin
            tags: [User]
            consumes:
                - application/json
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/definitions/SigninRequest"
            responses:
                "200":
                    description: Signin User using email and password
    /users/signup :
        post:
            summary: User signup
            tags: [User]
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            $ref: "#/components/schemas/User"
            responses:
                "200":
                    description: User was created successfully
    /users/recentProducts/{userId} :
        get:
            summary: Get the User's recently viewed products list
            tags: [User]
            consumes:
                - application/json
            parameters:
            - name: userId
              in: path
              required: true
              description: Autogenerated User ID by Mongoose
              schema:
                type: string
            responses:
                "200":
                    description: Recently viewed products by user
    /users/recentProducts/{userId}/{productId} :
        post:
            summary: Add a product to the User's recent products list
            tags: [User]
            consumes:
                - application/json
            parameters:
            - name: userId
              in: path
              required: true
              description: Autogenerated User ID by Mongoose
              schema:
                type: string
            - name: productId
              in: path
              required: true
              description: Autogenerated Product ID by Mongoose
              schema:
                type: string
            responses:
                "200":
                    description: Product Successfully Added to Recently Viewed Products

        delete:
            summary: Remove a product from the User's recent products list
            tags: [User]
            consumes:
                - application/json
            parameters:
            - name: userId
              in: path
              required: true
              description: Autogenerated User ID by Mongoose
              schema:
                type: string
            - name: productId
              in: path
              required: true
              description: Autogenerated Product ID by Mongoose
              schema:
                type: string
            responses:
                "200":
                    description: Recently viewed product deleted succesfully
    /users/favouriteBlogs/{userId} :
        get:
            summary: Get User's favourite blogs list
            tags: [User]
            consumes:
                - application/json
            parameters:
            - name: userId
              in: path
              required: true
              description: Autogenerated User ID by Mongoose
              schema:
                type: string
            responses:
                "200":
                    description: Favourite Blogs Successfully fetched
    /users/favouriteBlogs/{userId}/{blogId} :
        post:
            summary: Add a Blog to the User's Favorite blogs list
            tags: [User]
            consumes:
                - application/json
            parameters:
            - name: userId
              in: path
              required: true
              description: Autogenerated User ID by Mongoose
              schema:
                type: string
            - name: blogId
              in: path
              required: true
              description: Autogenerated Blog ID by Mongoose
              schema:
                type: string
            responses:
                "200":
                    description: Blog Successfully Added to favourite blogs list
        delete:
            summary: Remove a Blog from the User's Favorite blogs list
            tags: [User]
            consumes:
                - application/json
            parameters:
            - name: userId
              in: path
              required: true
              description: Autogenerated User ID by Mongoose
              schema:
                type: string
            - name: blogId
              in: path
              required: true
              description: Autogenerated Blog ID by Mongoose
              schema:
                type: string
            responses:
                "200":
                    description: Blog Successfully Removed from favourite blogs list
    /users/cartProducts/{userId} :
        get:
            summary: Get the Products in cart of the User
            tags: [User]
            consumes:
                - application/json
            parameters:
            - name: userId
              in: path
              required: true
              description: Autogenerated User ID by Mongoose
              schema:
                type: string
            responses:
                "200":
                    description: Get the Products in cart of the User
    /users/cartProducts/{userId}/{productId} :
        post:
            summary: Add a Blog to the User's Favorite blogs list
            tags: [User]
            consumes:
                - application/json
            parameters:
            - name: userId
              in: path
              required: true
              description: Autogenerated User ID by Mongoose
              schema:
                type: string
            - name: productId
              in: path
              required: true
              description: Autogenerated Product ID by Mongoose
              schema:
                type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/definitions/Cart"
            responses:
                "200":
                    description: Product Successfully Added to Cart

        delete:
            summary: Remove product from the cart
            tags: [User]
            consumes:
                - application/json
            parameters:
            - name: userId
              in: path
              required: true
              description: Autogenerated User ID by Mongoose
              schema:
                type: string
            - name: productId
              in: path
              required: true
              description: Autogenerated Product ID by Mongoose
              schema:
                type: string
            responses:
                "200":
                    description: Successfully Removed product from cart
components:
    schemas:
        User:
            type: object
            required:
                - firstName
                - lastName
                - email
                - password
                - DOB
                - file
            properties:
                firstName:
                    type: string
                    description: First Name of the User
                lastName:
                    type: number
                    description:  Last Name of the User
                email :
                    type: string
                    description: Email of the User
                password :
                    type: string
                    description: Password of User
                DOB :
                    type: string
                    description: Date of Birth of the User
                file:
                    type : file
                    description : Profile Image of the User
...
definitions:
  SigninRequest:
    type: object
    items:
      email: 
       type: string 
       required: true
       description: User Email
      password: 
       type: string 
       required: true
       description: User password 
...
  Cart:
    type: object
    required:
        - quantity
    properties:
        quantity:
            type: number
            description: Quantity of the Product
    example:
        quantity: 1